Creating vector of RealCards from xml and positioning virtcards around it
I also added a demo2.xml for testing, but the media will have to be
downloaded or it will crash. It is not pretty and xml is not handled in
a very good way. The interesting code is commented since it would be
harder to test some things with it and the scale is not good matched.